Adding licenses and renaming
