+++ src/soot/jimple/spark/internal/ClientAccessibilityOracle.java
@@ -1,0 +1,46 @@
+/* Soot - a J*va Optimization Framework
+ * Copyright (C) 2002 Ondrej Lhotak
+ *
+ * This library is free software; you can redistribute it and/or
+ * modify it under the terms of the GNU Lesser General Public
+ * License as published by the Free Software Foundation; either
+ * version 2.1 of the License, or (at your option) any later version.
+ *
+ * This library is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
+ * Lesser General Public License for more details.
+ *
+ * You should have received a copy of the GNU Lesser General Public
+ * License along with this library; if not, write to the
+ * Free Software Foundation, Inc., 59 Temple Place - Suite 330,
+ * Boston, MA 02111-1307, USA.
+ */
+
+package soot.jimple.spark.internal;
+
+import soot.SootField;
+import soot.SootMethod;
+
+/**
+ * The decision whether a libraries field or method is accessible for a client
+ * can be different for different analyses.
+ * 
+ * This interface provides methods to define how this decision will be made.
+ * 
+ * @author Florian Kuebler
+ *
+ */
+public interface ClientAccessibilityOracle {
+
+	/**
+	 * Determines whether the method is accessible for a potential library user.
+	 */
+	public boolean isAccessible(SootMethod method);
+
+	/**
+	 * Determines whether the field is accessible for a potential library user.
+	 */
+	public boolean isAccessible(SootField field);
+
+}
